Initial commit.
Add current state.  The user-side functions/macros are pretty ugly.
Function replacement with the usual SETF FDEFINITION stuff works,
accessing generic functions works as well, but is in an early stage.
Both parts are not merged in any way.  NOINLINE is probably necessary,
so implementation-specific ways to replace compiled references would be
a nice addition.  In any case testing with this should probably be done
on low optimisation settings to prevent surprises.